+++ libs/eslint-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @teo0098/eslint-config
 
+## 0.0.2
+
+### Patch Changes
+
+- Disable a rule for preferring a default export
+
+- Disable a rule for preferring function declaration over arrow function
+
 ## 0.0.1
 
 ### Patch Changes